refactor: 이전 도메인 따로 패키지로 관리 #264
